requirements.txt
numpy
pyscopg2
python-dotenv